+++ source/private/ConvertTo-StringData.ps1
@@ -1,0 +1,12 @@
+function ConvertToStringData {
+    [CmdletBinding()]
+    param(
+        [Parameter(ValueFromPipeline = $true)]
+        $InputObject
+    )
+    switch ($InputObject.GetType().FullName) {
+        'System.Collections.Hashtable' {
+            ($InputObject.Keys | % { "$_=$($InputObject.$_)" }) -join "`n"
+        }
+    }
+}